Move inventory to inventory tab and remove tokens sidebar
