+++ src/Header.jsx
@@ -10,7 +10,6 @@ import ioManager from '../io-manager.js'
 import { Character } from './components/general/character';
 import { CharacterSelect } from './components/general/character-select';
 import { Equipment } from './components/general/equipment';
-import { Tokens } from './tabs/tokens';
 import { registerIoEventHandler, unregisterIoEventHandler } from './components/general/io-handler';
 import { AppContext } from './components/app';
 import { AvatarIcon } from './AvatarIcon';
@@ -282,15 +281,6 @@ export default function Header() {
                         claims={claims}
                         panelsRef={panelsRef}
                     /> */}
-                </div>
-                <div className={styles.panels}>
-                    <Tokens
-                        nfts={nfts}
-                        hacks={hacks}
-                        address={address}
-                        setNfts={setNfts}
-                        loginFrom={loginFrom}
-                    />
                 </div>
             {/* </div> */}
         </div>
